:recycle: Refactored RequestResponseLoggingFilter
